Added "Stadium" features to battles.
Formats with Stadium Selection keep track of music and background selection,
and broadcast it to all with the '-stadium' minor action. Combatants can use
the new Battle Options menu to choose music/battlefields from a list of them.
It uses a new command '/stadium' behind the scenes. Stadium Selection is
applied to the TPPLA formats.

Also minor fix to stpplb. Also modified the Mix and Mega ruleset to include
the proper scripts (which hopefully works).